feat(l2): add support for web3signer (#2714)
**Motivation**

<!-- Why does this pull request exist? What are its goals? -->
Many operators will want to use a remote signer instead of having the
private keys on the same server as the sequencer.

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->
Replace all uses of a private key with a new `Signer` enum. This signer
can be either `Local` or `Remote` and can be lately extended. This aims
to standardise the way all kind of messages are signed across the L2 and
facilitate the setup via flags or environment

<!-- Link to issues: Resolves #111, Resolves #222 -->

---------

Co-authored-by: Javier Rodríguez Chatruc <49622509+jrchatruc@users.noreply.github.com>
Co-authored-by: fedacking <francisco.gauna@lambdaclass.com>
Co-authored-by: Avila Gastón <72628438+avilagaston9@users.noreply.github.com>