fix(l2): tdx to L2 prover client communication (#3599)
**Motivation**

The communication between the tdx prover client and the l2 prover server
is currently broken because the nix build does not set the
`VERGEN_GIT_SHA` variable.

**Description**
- Modify the ci to use run the image built with nix with QEMU, instead
of building the quote-gen as a binary and running it as a regular
binary.
- Add `gitRev` as a parameter to nix build
   - Use this parameter to set the env var `VERGEN_GIT_SHA`
- Add a check to make sure the length of the commit hash is the same one
that Vergen emits.
- Modify build.rs to get the value from the env var and if the value is
not set then it uses Vergen
- Remove build.rs files for zk provers and tdx quote-gen because the
values are added when building l2 crate.
- Successful run:
[here](https://github.com/lambdaclass/ethrex/actions/runs/16275888354/job/45954574993?pr=3599)

Closes #3587