fix(levm): preemptively resize memory before executing call (#3592)
**Motivation**
When executing a `CALL` opcode, a transfer might take place. In this
case the instruction does contain a return data offset and a return data
size but as we don't have return data to write into memory we don't
expand the memory.
This can cause problems with other opcodes later on (such as MSTORE,
MLOAD, etc) which calculate their gas cost based on the difference
between the current size of the memory and the new size, making them
more expensive as the memory will be smaller due to return data from
transfers not being accounted for.
This PR aims to solve this by preemptively resizing the memory before
executing the call, so that the memory gets expanded even if no return
data is written to it.
This bug was found on Sepolia transaction:
https://sepolia.etherscan.io/tx/0xa1765d420522a40d59d15f8dee1bf095499be687d6e1a7c978fc87eb85bce948
<!-- Why does this pull request exist? What are its goals? -->

**Description**
* Preemptively resize memory before executing a call in opcode `CALL`
<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->
**Questions**
Should this behaviour also apply to other call types?
Closes #issue_number