feat(levm): execute arbitrary bytecode (#3626)
**Motivation**

<!-- Why does this pull request exist? What are its goals? -->

**Description**

Have a runner for LEVM that expects some **inputs** like Transaction,
Fork, etc. in an `json` file and **bytecode in mnemonics** in another
file. Stack and memory can be preloaded within the `json`.
More info in the `README.md`

Sidenote: I had to do a refactor in LEVM setup because for me to be able
to alter the stack and memory before executing these have to be
initialized in the `new()`, thing that we weren't doing. So we now
initialize the first callframe there and not in `execute()`.

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #3583

---------

Co-authored-by: fmoletta <99273364+fmoletta@users.noreply.github.com>
Co-authored-by: Edgar <git@edgl.dev>