fix(levm): memory bug when storing data (#3774)
**Motivation**

<!-- Why does this pull request exist? What are its goals? -->

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->
We didn't realize that #3564 introduced a bug when storing data of
length zero. This aims to fix it.
I also delete a resize check that's completely unnecessary

Tested the fix and it works. I now am able to execute the blocks
mentioned in the issue of this PR without any problems at all.

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #3775