chore(l1): change logs in hive to info by default (#3767)
**Motivation**

In the PR #2975 the default value for the `make run-hive` was changed to
error. I propose changing this to info (3), as we usually run make hive
to try to see a problem with the test. For the CI I propose we change it
to log level error (1), as we can't actually look at those logs.

**Description**

- Changed makefile `SIM_LOG_LEVEL` default value to 3 (info)
- Added to the ci workflows `--sim.loglevel 1` which corresponds to
error.