refactor(l2): improve naming and standardize arguments in l2 tests (#3790)
**Motivation**

Currently the L2 tests:
* use unintuitive names (eth_client vs proposer_client, meaning l1 and
l2)
* do not have a consistent ordering of parameters
* are inconsistent on when things (bridge address and rich private key)
are given as parameter vs obtained from a function

**Description**

This PR improves that, and gets the "noisy" changes out of the way for
further improvements.

The rich private key was kept as a parameter to allow giving different
ones (in the future, this would allow parallelizing the tests). The
bridge address now always uses the function, since it won't change in
the middle of the test.