feat(l2): show peer count in monitor when in based mode (#3668)
**Motivation**

Currently peer count is a placeholder ('NaN') so we want an actual
implementations.

The peer count of the L2 should only be shown in based bode, because in
other (non-based) modes peers aren't _needed_ and having more isn't
necessarily better.

**Description**

Adds a EthClient method to query net_peerCount, and uses it to get the
peer count from the local node. Only displays the value if based mode is
enabled.

Closes #3516

---------

Co-authored-by: Ivan Litteri <67517699+ilitteri@users.noreply.github.com>