feat(l2): add privileged transaction cap (#3776)
**Motivation**

If the l2 batches include too many privileged transactions, it may
happen that the gas cost of `getPendingTransactionsVersionedHash(uint16
n)` becomes so high it's impossible to commit batches.

**Description**

To prevent this, we record the privileged transaction count of the batch
being built and ensure no more than a fixed amount (300 for now) is
added.

We add a spammer (that makes 1800 privileged transactions) as a
regression test.

Additionally, this allows enforcing intra-batch ordering of privileged
transactions. This prevents the following scenario from happening:
* A sends two deposits with gas values 1 (deposit 1) and 10M (deposit 2)
* B then sends a deposit with gas value 1 (deposit 3)
* The sequencer includes A's first deposit, but not the second because
it's too big for the block. Deposit 3 is included because it's small.
* The resulting inclusion order is 1, 3. This out-of-order and therefore
the batch is invalid.