test(l2): monitor withdrawal widget integration tests (#3681)
**Motivation**

We want to test the logic of the withdrawal widget, which is the only
one with significant logic behind it.

As for the other widgets:

* NodeStatus/Batch are hard to test because they depend on
out-of-control events, and don't to much beyond calling RPCs.
* Mempool is hard to test, because we don't have much control (aside
from "tricks" such as setting high nonces) over when tranactions are
pulled from the mempool. Also it doesn't have much logic inside. It also
relies on being part of the sequencer.
* Deposits do have a little bit more logic, and can be somewhat
controlled (deposits shouldn't magically disappear if nothing else is
sending transactions). However, it relies on having access to the
sequencer store so it's deferred.

**Description**

Both ETH and ERC20 withdrawals are tested, as part of other integration
tests.

Some changes are made to make testing easier, and a bug (incorrect topic
index) is fixed.

Closes #3526

---------

Co-authored-by: Ivan Litteri <67517699+ilitteri@users.noreply.github.com>