fix(l2): integration tests hotfix (#3892)
**Motivation**

after merging #3776 a test was added that spams the bridge with big gas
limit transactions. In #3851 we upped the gas limit of all transactions
by x2 because the estimation was too low. These two contributed to
exceed the block gas limit.

This PR lowers the amount of spam txs

thanks @iovoid !