perf(l1): batch new pooled tx hashes (#3912)
**Motivation**

Reduce throughput degradation seen in localnet scenarios with more than
one ethrex

**Description**

Up until now we were sending individual messages for each tx we haven't
shared with the network. We discovered [in this
branch](https://github.com/lambdaclass/ethrex/pull/3900) that removing
the send of new pooled transactions along with the broadcast of
transactions removed the issue.

Batching the new pooled transactions in localnets with two ethrex show
metrics similar to previous runs for localnets with only one ethrex.

Apart from this a Test have been disable due to being a false positive
previously, here is an issue explaining it: #3914