chore(l1): preventing errors to propagate to spawned (#3840)
**Motivation**

Error lines in spawned logs [were
muted](https://github.com/lambdaclass/ethrex/pull/3467) because there
was a lot of spam regarding failed connections and handshakes. That was
not safe as other important problems may get muted too.

**Description**

This PR handles the connection errors on handshake and initialization
and logs a lower priority log line (debug) as it is pretty common.

It also reverts the log directive to mute spawned initialization errors.

it requires [spawned
0.2.2](https://github.com/lambdaclass/spawned/pull/39).

Closes #3472