fix(levm): fix tracing delegatecall (#3823)
**Motivation**

- The from in the DELEGATECALL trace was wrong.

**Description**

I wanted the trace of a transaction that made use of DELEGATECALL and I
saw that ethrex had a weird response, so I tried getting the same trace
with geth and the response was what I expected it to be. So that's how I
realized that we had a bug :)

[Hoodi
transaction](https://hoodi.etherscan.io/tx/0x0bca10d4b752d3c630f4666a226a37d57a08b0b042b18888dbcb34de73679cf0)
```
curl <rpc_hoodi_node> \
-X POST \
-H "Content-Type: application/json" \
--data '{"method":"debug_traceTransaction","params":["0x0bca10d4b752d3c630f4666a226a37d57a08b0b042b18888dbcb34de73679cf0", {"tracer": "callTracer"}], "id":1,"jsonrpc":"2.0"}'
```

Basically if our trace was used in a block explorer (be it blockscout or
etherscan) the delegatecall "From" field would be the EOA and that is
wrong! Check the actual tracing in etherscan
[here](https://hoodi.etherscan.io/tx/0x0bca10d4b752d3c630f4666a226a37d57a08b0b042b18888dbcb34de73679cf0#internal)