fix(levm): ignore offset value when size is zero (#3950)
**Motivation**
We recently ran into a gas mismatch when running full sync in sepolia
testnet. The problem was that we were reverting earlier due to an
`ExceptionHalt::VeryLargeNumber`. This happened during a `STATICCALL`
opcode execution where the `arg_offset` value taken from the stack was a
very large number. Upon comparing our implementation with geth we
realized that the difference lies in that we do ignore the value of the
offset when calculating the memory size If the size is zero, but by that
time we have already converted the offset into a usize value, and failed
if we couldn't do so.
The transaction where this bug was discovered is
https://sepolia.etherscan.io/tx/0x0495f636d6d69c91e40afc44d3bd12800831d848391666cf23325720fd9ebfd4
This PR solves this problem and aims to solve future related problems by
adding a utility function `size_offset_to_usize` which first converts
the size into a size value and if it is not zero then converts the
offset into a size value, if the size is zero, offset will be ignored
and given a default value as it won't be used.
It also gets rid of some explicit error handling by adding utility
function `u256_to_usize` to checkedly convert U256 to usize. Some tweaks
have also been performed while checking against geth handling.
<!-- Why does this pull request exist? What are its goals? -->

**Description**
* Add util  function `u256_to_usize` which converts a U256 to usize.
* Add util function `size_offset_to_usize` which converts U256 size and
offset to usize, ignoring the offset value if size is zero.
<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes None, but is needed in order to progress through #1676