refactor(l1): tooling/sync make targets to be generic across networks (#3869)
**Motivation**
The current Makefile on tooling/sync has a bunch of different yet very
similar targets that vary depending on whether they are used on hoodi or
holesky networks. This is not much of an issue with just two networks,
but would become quite bloated if we wanted to also add targets for
sepolia and mainnet networks.
The aim of this PR is to refactor these duplicated targets into a single
one leveraging the `NETWORK` env var and also allow these targets to
support mainnet and sepolia networks
<!-- Why does this pull request exist? What are its goals? -->

**Description**
* Replace inner targets `holesky-lighthouse` & `hoodi-lighthouse` with
generic `start-lighthouse` target that depends on `NETWORK` env var
* Replace inner targets `start-ethrex-holesky` & `start-ethrex-hoodi`
with generic `start-ethrex` target that depends on `NETWORK` env var
* Replace targets `start-holesky-metrics-docker` &
`start-hoodi-metrics-docker` with generic `start-metrics-docker` target
that depends on `NETWORK` env var + A wildcard target
`start-%-metrics-docker` that sets the `NETWORK` env var to the wildcard
so that previous targets are still functional.
* Replace targets `flamegraph-holesky` & `flamegraph-hoodi` with generic
`flamegraph-inner` target (as `flamegraph` is already its own target)
that depends on `NETWORK` env var + A wildcard target `flamegraph-%r`
that sets the `NETWORK` env var to the wildcard so that previous targets
are still functional.
* `NETWORK` env var is no longer strictly required, and has default
value set to mainnet (Docs updated)
* Removed check for `NETWORK` env var being present (as it now uses
default value)
* Updated handling of BOOTNODES:
* Removed `HOLESKY_BOOTNODES` and `HOODI_BOOTNODES` as they are already
present on the preset bootnodes file
* Only set the `--bootnodes` flag when running ethrex if the `BOOTNODES`
env var was set
* (Misc) Fix `help` target so that it also includes wildcard targets +
set it as default target
* (Misc) Mention installing `flamegraph` crate on README`
<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #3838 (No longer needed with this change)

---------

Co-authored-by: Mario Rugiero <mrugiero@gmail.com>