chore(l1): cache tx hash (#4049)
**Motivation**

We saw that the hashing of txs takes a significant amount of time and
part of it is repeated.

**Description**

This PR create a new public hash function that allows to cache the
computation of the hash for subsequent calls.

After merging main there are some more changes needed. I'm taking a look
at them.

Closes #issue_number