refactor(l1): `ethrex_replay` use rkyv for encoding caches (#4169)
**Motivation**

We want to change the encoding for the caches that the replay generates
when executing or proving a block from a JSON pretty formatted to one
serialized using `rkyv`.

**Description**

* Switched cache serialization from JSON to binary using the `rkyv`
library , including changes to `load_cache` and `write_cache` to use
`rkyv` for reading and writing cache files
* Updated `Cache` and `L2Fields` structs to derive `Archive`,
`RSerialize`, and `RDeserialize` traits for compatibility with `rkyv`.
* Changed cache file extensions from `.json` to `.bin` in all relevant
cache-related functions.
* Added `rkyv.workspace = true` to `cmd/ethrex_replay/Cargo.toml` to
manage the `rkyv` dependency at the workspace level.