feat(l2): add support for multiple networks on a single prover (#4186)
**Motivation**

<!-- Why does this pull request exist? What are its goals? -->
Prover may be idle for a long time if batch generation frequency is low,
wasting resources.

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->
Allow the prover to have multiple proof coordinator endpoints. This can
be used both for multiple networks as for redundancy. Since the prover
is stateless, it doesn't matter if the endpoints are from different
networks.
The prover will try to get a task sequentially on each endpoint. After
all endpoints are checked and proofs generated (in case there's anyone
to generate), the prover will sleep `--proving-time` milliseconds and
start again

<!-- Link to issues: Resolves #111, Resolves #222 -->

---------

Co-authored-by: LeanSerra <46695152+LeanSerra@users.noreply.github.com>