feat(l1): added fusaka support for runners (#4163)
**Motivation**

Adding the boilerplate in the runners to be able to execute the Fusaka
tests in the future when we start implementing it

**Description**

This PR changes the fixture URLs to the most recent execution spec tests
release that supports Fusaka. Additionally, it adds the parameters
needed to the structs the runners use to be able to parse and execute
the test; although at the moment they're skipped by default. The Osaka
tests were also added to the Hive daily report.

Closes #4114